prodbuild 017

+++ console/alunos/editaluno/rmedit.php
@@ -6,7 +6,11 @@
   $id     = $_POST['id'];
   $classeRaw = $_POST['classe'];
 
-  $classe = $classeRaw.' Classe'; // like 1a Classe, 2a Classe, 3a Cl...
+  if ($classeRaw !== "Pre") {
+    $classe = $classeRaw.' Classe'; // like 1a Classe, 2a Classe, 3a Cl...
+  } else {
+    $classe = "Pré-Escolar";
+  }
 
   # Define all classes
   $disciplinas_1a = ["port_1a", "mat_1a", "ing_1a", "artesvisuais_1a", "musica_1a", "edf_1a", "danca_1a"];
@@ -15,15 +19,17 @@
   $disciplinas_4a = ["port_4a", "mat_4a", "ing_4a", "artesvisuais_4a", "musica_4a", "edf_4a", "danca_4a", "cnaturais_4a", "csociais_4a"];
 
   # Delete the profile record
-  $sql = "DELETE FROM profiles WHERE id='".$id."'";
-  mysqli_query($conn, $sql);
+  $red_sql = "DELETE FROM profiles WHERE id='".$id."'";
+  mysqli_query($conn, $red_sql);
 
   # Delete the records from disciplinas
-  foreach (${"disciplinas_".$classeRaw} as $disciplina) {
-    // $query3 = "INSERT INTO `$disciplina`(nome_id, year) SELECT id, started_year FROM profiles, temp_profiles WHERE profiles.username=temp_profiles.username";
-    $sql = "DELETE FROM `$disciplina` WHERE nome_id='".$id."'";
-    $red_sql = mysqli_query($conn, $sql);
-  };
+  if ($classeRaw !== "Pre") {
+    foreach (${"disciplinas_".$classeRaw} as $disciplina) {
+      // $query3 = "INSERT INTO `$disciplina`(nome_id, year) SELECT id, started_year FROM profiles, temp_profiles WHERE profiles.username=temp_profiles.username";
+      $sql = "DELETE FROM `$disciplina` WHERE nome_id='".$id."'";
+      $red_sql = mysqli_query($conn, $sql);
+    }
+  }
 
   # Check request result:
   if (!$red_sql){
@@ -32,9 +38,9 @@
 
   } else {
     # if everything went smoothly :P
-    echo "done";
+    echo $id;
   }
 
-  $conn->close();
+  // $conn->close();
 
 ?>
